Refactor SSL configuration
Refactors the way this role configures SSL and nginx virtual host. By
default, it will configure munki for delivery over HTTP (without SSL)
unless the user provides the path to the SSL certificate and private
key, in which case it will install and set up HTTPS. This should make
the configuration of SSL simpler and more straightforward and less
dependant on how the certificate is installed on the system.